services/alert-engine-service/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "alert-engine-service"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "confluent-kafka>=2.3.0",
    "pydantic>=2.0",
    "psycopg2-binary>=2.9",
]
